pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mindlayer-ai"
version = "0.1.0"
description = "Open-source, model-agnostic memory layer for LLMs"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.9"
keywords = ["llm", "memory", "ai", "chatbot", "rag"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = []

[project.optional-dependencies]
vector = [
    "sqlite-vec>=0.1.0",
    "fastembed>=0.3.0",
]
llm = [
    "llama-cpp-python>=0.2.0",
    "huggingface-hub>=0.20.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
]

[project.urls]
Homepage = "https://github.com/Genious07/mindlayer"
Repository = "https://github.com/Genious07/mindlayer"
Issues = "https://github.com/Genious07/mindlayer/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["mindlayer*"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
